requirements.txt
python-telegram-bot[job-queue]==20.6
speechrecognition==3.9.0
pydub==0.25.1
gtts==2.3.1
together==1.4.6  # Убедитесь, что версия актуальна
